+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v0.7.0 (in development)
 - `parse()` now accepts strings as input
 - `parse()` now returns a generator of custom objects instead of triples of
   strings
+- Gave `PropertiesFile` a settable `timestamp` property
 
 v0.6.0 (2020-02-28)
 -------------------