+++ docs/changelog.rst
@@ -8,6 +8,7 @@ v0.7.0 (in development)
 - `parse()` now accepts strings as input
 - `parse()` now returns a generator of custom objects instead of triples of
   strings
+- Gave `PropertiesFile` a settable `~PropertiesFile.timestamp` property
 
 v0.6.0 (2020-02-28)
 -------------------